deps
wasm'
'

change sthat i dont care

small fn

mod: removed unnecessary mod

mod: removed unnecessary mod

js: proper listner handlel'

update ingore file

add: added tests and update the version
