+++ .gitignore
@@ -1,3 +1,2 @@
 /target
 .direnv/
-templates/temp.html
