--- static/js/editor/prev_editor.js
@@ -1,8 +1,0 @@
-document.addEventListener("DOMContentLoaded", function() {
-   const savedEditor = localStorage.getItem("editor");
-   if (savedEditor === "vscode") {
-      vimModeEnabled = true;
-   } else if (savedEditor === "vim") {
-      vimModeEnabled = false;
-   }
-});
